# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Versatile Timesheet Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees with pagination, sorting, and filtering
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: department
          in: query
          schema: { type: string }
        - name: managerId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeList"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/bulk-upload:
    post:
      operationId: bulkUploadEmployees
      tags: [employees]
      summary: Bulk upload employees from Excel/CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkEmployeeUpload"
      responses:
        "200":
          description: Upload result with row-wise errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUploadResult"
  /employees/{employeeId}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete (deactivate) employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
  /employees/{employeeId}/profile:
    get:
      operationId: getEmployeeProfile
      tags: [employees]
      summary: Get employee profile with metrics
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Employee profile with metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeProfile"
  /employees/{employeeId}/direct-reports:
    get:
      operationId: getDirectReports
      tags: [employees]
      summary: Get direct reports of a manager
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of direct reports with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerProfile"

  # CLIENTS
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientList"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: clientId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: clientId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      parameters:
        - name: clientId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted

  # PROJECTS
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activity categories
      responses:
        "200":
          description: List of activity categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create activity category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{activityId}:
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update activity category
      parameters:
        - name: activityId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete activity category
      parameters:
        - name: activityId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted

  # TIMESHEETS
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent timesheet activity feed
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityFeedItem"
...
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        designation:
          type: string
        role:
          type: string
          enum: [Employee, Manager, Admin]
        managerId:
          type: ["string", "null"]
        managerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [Active, Inactive]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, employeeId, name, email, department, designation, role, status, createdAt, updatedAt]
    EmployeeInput:
      type: object
      properties:
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        designation:
          type: string
        role:
          type: string
          enum: [Employee, Manager, Admin]
        managerId:
          type: ["string", "null"]
        status:
          type: string
          enum: [Active, Inactive]
      required: [employeeId, name, email, department, designation, role, status]
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
        designation:
          type: string
        role:
          type: string
          enum: [Employee, Manager, Admin]
        managerId:
          type: ["string", "null"]
        status:
          type: string
          enum: [Active, Inactive]
    EmployeeList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required: [data, total, page, pageSize]
    EmployeeProfile:
      type: object
      properties:
        employee:
          $ref: "#/components/schemas/Employee"
        metrics:
          $ref: "#/components/schemas/EmployeeMetrics"
      required: [employee, metrics]
    EmployeeMetrics:
      type: object
      properties:
        totalSubmitted:
          type: integer
        pending:
          type: integer
        rejected:
          type: integer
        approved:
          type: integer
        approvalRate:
          type: number
      required: [totalSubmitted, pending, rejected, approved, approvalRate]
    ManagerProfile:
      type: object
      properties:
        manager:
          $ref: "#/components/schemas/Employee"
        metrics:
          $ref: "#/components/schemas/ManagerMetrics"
        directReports:
          type: array
          items:
            $ref: "#/components/schemas/DirectReportSummary"
      required: [manager, metrics, directReports]
    ManagerMetrics:
      type: object
      properties:
        totalReports:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        pendingApprovals:
          type: integer
      required: [totalReports, approved, rejected, pendingApprovals]
    DirectReportSummary:
      type: object
      properties:
        employee:
          $ref: "#/components/schemas/Employee"
        submitted:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
      required: [employee, submitted, approved, rejected]
    BulkEmployeeUpload:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/BulkEmployeeRow"
      required: [rows]
    BulkEmployeeRow:
      type: object
      properties:
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        designation:
          type: string
        role:
          type: string
        managerId:
          type: string
        status:
          type: string
      required: [employeeId, name, email, department, designation, role, status]
    BulkUploadResult:
      type: object
      properties:
        totalRows:
          type: integer
        successCount:
          type: integer
        errorCount:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/BulkUploadError"
      required: [totalRows, successCount, errorCount, errors]
    BulkUploadError:
      type: object
      properties:
        row:
          type: integer
        field:
          type: string
        message:
          type: string
      required: [row, field, message]
    Client:
      type: object
      properties:
        id:
          type: string
        clientCode:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientCode, name, status, createdAt, updatedAt]
    ClientInput:
      type: object
      properties:
        clientCode:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
      required: [clientCode, name, status]
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
    ClientList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required: [data, total, page, pageSize]
...
    Activity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
        createdAt:
          type: string
      required: [id, name, status, createdAt]
    ActivityInput:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
      required: [name, status]
    ActivityUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
...
    ActivityFeedItem:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
        employeeName:
          type: string
        weekStartDate:
          type: string
        timestamp:
          type: string
        timesheetId:
          type: string
      required: [id, action, employeeName, weekStartDate, timestamp, timesheetId]
...